+++ src/buttons/all_buttons.hpp
@@ -1,5 +1,5 @@
 #pragma once
-#include <iostream>
+
 #include <raylib.h>
 #include "button.hpp"
