+++ src/cells/gold.cpp
