+++ src/cells/gold.hpp
