+++ src/cells/grid_cell.cpp
