+++ src/grid/map_object.hpp
