+++ src/main.cpp
@@ -1,4 +1,4 @@
-#include "main_window.hpp"
+#include "windows/main_window.hpp"
 
 Vector2 mouse_position = { 100.f, 100.f };
 window_type draw_window = window_type::TOWN;
