+++ src/windows/main_window.hpp
@@ -3,8 +3,6 @@
 #include <raylib.h>
 
 #include "window.hpp"
-#include "grid.hpp"
-#include "dwarf.hpp"
 #include "town.hpp"
 #include "shop.hpp"
 #include "mine.hpp"
